requirements.txt
langchain>=0.1.16
langchain-community>=0.0.34
langchain-openai>=0.1.6
faiss-cpu>=1.7.4
chainlit>=1.0.0
pymupdf>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.1
pillow>=10.3.0
openai>=1.30.1 